Rename show-line-of-sight-between-geoelements to show-exploratory-line-of-sight-between-geoelements
