+++ samples/show-exploratory-line-of-sight-between-geoelements/show-exploratory-line-of-sight-between-geoelements.png
