+++ samples/show-exploratory-line-of-sight-between-geoelements/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
         <activity
             android:exported="true"
             android:name=".DownloadActivity"
-            android:label="@string/show_line_of_sight_between_geoelements_app_name">
+            android:label="@string/show_exploratory_line_of_sight_between_geoelements_app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
             </intent-filter>
@@ -15,7 +15,7 @@
         <activity
             android:exported="true"
             android:name=".MainActivity"
-            android:label="@string/show_line_of_sight_between_geoelements_app_name">
+            android:label="@string/show_exploratory_line_of_sight_between_geoelements_app_name">
         </activity>
     </application>
