--- samples/show-line-of-sight-between-geoelements/README.metadata.json
@@ -1,33 +1,0 @@
-{
-    "category": "Analysis",
-    "description": "Show an exploratory line of sight between two moving objects.",
-    "formal_name": "ShowLineOfSightBetweenGeoelements",
-    "ignore": false,
-    "images": [
-        "show-line-of-sight-between-geoelements.png"
-    ],
-    "keywords": [
-        "3D",
-        "geoview-compose",
-        "line of sight",
-        "visibility",
-        "visibility analysis",
-        "AnalysisOverlay",
-        "ExploratoryGeoElementLineOfSight",
-        "ExploratoryLineOfSightTargetVisibility"
-    ],
-    "language": "kotlin",
-    "redirect_from": "",
-    "relevant_apis": [
-        "AnalysisOverlay",
-        "ExploratoryGeoElementLineOfSight",
-        "ExploratoryLineOfSightTargetVisibility"
-    ],
-    "snippets": [
-        "src/main/java/com/esri/arcgismaps/sample/showlineofsightbetweengeoelements/components/SceneViewModel.kt",
-        "src/main/java/com/esri/arcgismaps/sample/showlineofsightbetweengeoelements/DownloadActivity.kt",
-        "src/main/java/com/esri/arcgismaps/sample/showlineofsightbetweengeoelements/MainActivity.kt",
-        "src/main/java/com/esri/arcgismaps/sample/showlineofsightbetweengeoelements/screens/MainScreen.kt"
-    ],
-    "title": "Show line of sight between geoelements"
-}
